# Examples/APIBackend/static/swagger.yml
openapi: 3.0.1
info:
  title: ZenKey Example Backend API
  description: ZenKey Example Backend API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: /
paths:
  /auth/zenkey-signin:
    post:
      tags:
        - Auth
      summary: Complete ZenKey sign in to receive a token
      description: |
        Use this endpoint to finish the ZenKey sign in get an access token. Behind the scenes, the API server will exchange the ZenKey auth code for a ZenKey token and make a request to ZenKey for the user's info.
      operationId: sign-in
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            examples:
              minimal:
                $ref: '#/components/examples/MinimalSignInRequest'
              full:
                $ref: '#/components/examples/FullSignInRequest'
            schema:
              $ref: '#/components/schemas/SignInRequest'
          application/x-www-form-urlencoded:
            examples:
              minimal:
                $ref: '#/components/examples/MinimalSignInRequest'
              full:
                $ref: '#/components/examples/FullSignInRequest'
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: User does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDoesNotExistResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/token:
    post:
      summary: Refresh token (not fully implemented)
      description: |
        Use a refresh token to get a new access token
      operationId: refresh-token
      tags:
        - Auth
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - grant_type
                - refresh_token
              properties:
                grant_type:
                  type: string
                  description: Requested grant type. Only "refresh_token" is allowed
                  example: refresh_token
                refresh_token:
                  type: string
                  description: The refresh token to be exchanged for a new token
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Log out (not fully implemented)
      description: End a session by invalidating the access token to log the user out
      operationId: end-session
      tags:
        - Auth
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        200:
          description: Success
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  examples:
    MinimalSignInRequest:
      summary: A simple request
      value:
        client_id: my-client-id
        code: DZYeUmmQjKfTsZeRa6PukTps
        mccmnc: "311480"
        redirect_uri: https://mydomain.com/zenkey/callback
    FullSignInRequest:
      summary: A request with all the parameters
      value:
        client_id: my-client-id
        code: DZYeUmmQjKfTsZeRa6PukTps
        mccmnc: "311480"
        redirect_uri: https://mydomain.com/zenkey/callback
        code_verifier: ZYeUmmQjKfTsZeRa6PukTp
        correlation_id: 10001
        nonce: mmQjKfTsZeRa6P
        context: "Transfer $10.00 to Jesse"
        acr_values: "a1 a3"
...
  schemas:
    SignInRequest:
      required:
        - client_id
        - code
        - mccmnc
        - redirect_uri
      properties:
        client_id:
          type: string
          description: Client ID
          example: my-client-id
        code:
          type: string
          description: code from ZenKey SDK AuthorizedResponse
          example: DZYeUmmQjKfTsZeRa6PukTps
        redirect_uri:
          type: string
          description: Redirect URL
          example: https://mydomain.com/zenkey/callback
        code_verifier:
          type: string
          description: codeVerifier from ZenKey SDK AuthorizedResponse
        mccmnc:
          type: string
          description: MCCMNC of the user's carrier
          example: "311480"
        correlation_id:
          type: string
          description: the same correlation ID passed in the auth request.
            Used to correlate requests
        nonce:
          type: string
          description: the nonce passed in the auth request. If present, the
            backend will validate the nonce returned in the ID token to make
            sure it matches
        context:
          type: string
          description: the context passed in the auth request. If present,
            the backend will validate the context returned in the ID token
            to make sure it matches
          example: "Transfer $100 to Jesse?"
        acr_values:
          type: string
          description: the ACR values passed in the auth request. If present,
            the backend will validate the ACR values returned in the ID token
            to make sure they match. This should be a series of strings separated by spaces.
          example: "a1 a3"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
...
    UserDoesNotExistResponse:
      type: object
      properties:
        zenkey_attributes:
          $ref: '#/components/schemas/ZenKeyAttributes'
        error:
          type: string
        error_description:
          type: string
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: a JWT linked to the session. This JWT acts as an access token and must be passed in the Authorization header for authenticated requests made to the API backend. This is an internal token, not a token from ZenKey.
        refresh_token:
          type: string
          description: a refresh token to be used to renew the token. This is an internal token, not a token from ZenKey.
        token_type:
          type: string
          description: the type of the token
          example: "bearer"
        expires:
          type: number
          description: seconds remaining until the token expires
          example: 3600
    ZenKeyAttributes:
      type: object
      required:
        - sub
      properties:
        sub:
          type: string
          description: Unique ZenKey account identifier for the user
          example: my-zenkey-sub
        name:
          type: object
          properties:
            value:
              type: string
              description: User's full name
              example: Jesse Doe
            given_name:
              type: string
              description: User's first name
              example: Jesse
            family_name:
              type: string
              description: User's last name
              example: Doe
            verified:
              type: number
              example: 1
        phone_number:
          type: object
          properties:
            value:
              type: string
              description: User's phone number
              example: 555-555-5555
            verified:
              type: number
              example: 1
        postal_code:
          type: object
          properties:
            value:
              type: string
              description: User's postal code
              example: 55155
            verified:
              type: number
              example: 1
        email:
          type: object
          properties:
            value:
              type: string
              description: User's email address
              example: jesse@myzenkey.com
            verified:
              type: number
              example: 1
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: X-API-KEY